:recycle: refactor: 챌린지 코스 조회 쿼리 Challenge Repository로 이동
