+++ src/main/java/com/saisai/domain/challenge/repository/ChallengeRepositoryImpl.java
@@ -7,14 +7,19 @@ import static com.saisai.domain.reward.entity.QRewardEvent.rewardEvent;
 import static com.saisai.domain.ride.entity.QRide.ride;
 
 import com.querydsl.jpa.JPAExpressions;
+import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.saisai.domain.challenge.dto.projection.ChallengeCourseProjection;
+import com.saisai.domain.challenge.dto.projection.QChallengeCourseProjection;
 import com.saisai.domain.challenge.entity.ChallengeStatus;
-import com.saisai.domain.course.dto.projection.ChallengeCourseProjection;
-import com.saisai.domain.course.dto.projection.QChallengeCourseProjection;
+import com.saisai.domain.course.constant.CourseSortOption;
 import com.saisai.domain.reward.dto.projection.QRewardEventProjection;
 import com.saisai.domain.reward.entity.EventStatus;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.support.PageableExecutionUtils;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -22,6 +27,70 @@ import org.springframework.stereotype.Repository;
 public class ChallengeRepositoryImpl implements ChallengeRepositoryCustom {
 
     private final JPAQueryFactory queryFactory;
+
+    // 챌린지 코스 조회
+    @Override
+    public Page<ChallengeCourseProjection> findChallengeCourses(Pageable pageable,
+        CourseSortOption sortOption, Long userId) {
+
+        List<ChallengeCourseProjection> content = queryFactory
+            .select(new QChallengeCourseProjection(
+                course.id,
+                course.name,
+                course.level,
+                course.distance,
+                course.estimatedTime,
+                course.sigun,
+                course.image,
+                ride.count().coalesce(0L),
+                JPAExpressions
+                    .selectOne()
+                    .from(courseBookmark)
+                    .where(courseBookmark.user.id.eq(userId)
+                        .and(courseBookmark.course.id.eq(course.id))
+                    )
+                    .exists(),
+                challenge.status,
+                challenge.endedAt,
+                new QRewardEventProjection(
+                    rewardEvent.id,
+                    rewardEvent.status,
+                    rewardEvent.type,
+                    rewardEvent.value
+                )
+            ))
+            .from(challenge)
+            .innerJoin(challenge.course, course)
+            .leftJoin(ride).on(ride.course.id.eq(course.id))
+            .leftJoin(rewardEvent).on(
+                rewardEvent.challenge.id.eq(challenge.id)
+                    .and(rewardEvent.status.eq(EventStatus.ACTIVE)))
+            .where(course.isDeleted.eq(false)
+                .and(challenge.status.eq(ChallengeStatus.ONGOING)))
+            .groupBy(course.id, course.name, course.level, course.distance,
+                course.estimatedTime, course.sigun, course.image,
+                challenge.status, challenge.endedAt,
+                rewardEvent.id, rewardEvent.status, rewardEvent.type, rewardEvent.value)
+            .orderBy(sortOption.toOrderSpecifier(),
+                course.name.asc())
+            .offset(pageable.getOffset())
+            .limit(pageable.getPageSize())
+            .fetch();
+
+        JPAQuery<Long> total = queryFactory
+            .select(course.countDistinct())
+            .from(challenge)
+            .innerJoin(challenge.course, course)
+            .leftJoin(ride).on(ride.course.id.eq(course.id))
+            .leftJoin(rewardEvent).on(
+                rewardEvent.challenge.id.eq(challenge.id)
+                    .and(rewardEvent.status.eq(EventStatus.ACTIVE))
+            )
+            .where(course.isDeleted.eq(false)
+                .and(challenge.status.eq(ChallengeStatus.ONGOING)));
+
+        return PageableExecutionUtils.getPage(content, pageable, total::fetchOne);
+    }
 
     // 챌린지 진행 중인 코스 중에서 참가자가 많은 순으로 정렬 후 반환하는 메서드
     @Override
