+++ src/main/java/com/saisai/domain/course/repository/CourseRepositoryImpl.java
@@ -12,13 +12,10 @@ import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.saisai.domain.challenge.entity.ChallengeStatus;
 import com.saisai.domain.course.constant.CourseSortOption;
-import com.saisai.domain.course.dto.projection.ChallengeCourseProjection;
 import com.saisai.domain.course.dto.projection.CourseDetailsProjection;
 import com.saisai.domain.course.dto.projection.GeneralCourseProjection;
-import com.saisai.domain.course.dto.projection.QChallengeCourseProjection;
 import com.saisai.domain.course.dto.projection.QCourseDetailsProjection;
 import com.saisai.domain.course.dto.projection.QGeneralCourseProjection;
-import com.saisai.domain.reward.dto.projection.QRewardEventProjection;
 import com.saisai.domain.reward.entity.EventStatus;
 import java.util.List;
 import java.util.Optional;
@@ -83,64 +80,6 @@ public class CourseRepositoryImpl implements CourseRepositoryCustom {
             .leftJoin(ride).on(ride.course.id.eq(course.id))
             .where(course.isDeleted.eq(false)
                 .and(challenge.id.isNull()));
-
-        return PageableExecutionUtils.getPage(content, pageable, total::fetchOne);
-    }
-
-    // 챌린지 코스 조회
-    @Override
-    public Page<ChallengeCourseProjection> findChallengeCourses(Pageable pageable,
-        CourseSortOption sortOption, Long userId) {
-
-        List<ChallengeCourseProjection> content = queryFactory
-            .select(new QChallengeCourseProjection(
-                course.id,
-                course.name,
-                course.level,
-                course.distance,
-                course.estimatedTime,
-                course.sigun,
-                course.image,
-                ride.count().coalesce(0L),
-                JPAExpressions
-                    .selectOne()
-                    .from(courseBookmark)
-                    .where(courseBookmark.user.id.eq(userId)
-                        .and(courseBookmark.course.id.eq(course.id))
-                    )
-                    .exists(),
-                challenge.status,
-                challenge.endedAt,
-                new QRewardEventProjection(
-                    rewardEvent.id,
-                    rewardEvent.status,
-                    rewardEvent.type,
-                    rewardEvent.value
-                )
-            ))
-            .from(challenge)
-            .innerJoin(challenge.course, course)
-            .leftJoin(ride).on(ride.course.id.eq(course.id))
-            .leftJoin(rewardEvent).on(
-                rewardEvent.challenge.id.eq(challenge.id)
-                    .and(rewardEvent.status.eq(EventStatus.ACTIVE)))
-            .where(course.isDeleted.eq(false)
-                .and(challenge.status.eq(ChallengeStatus.ONGOING)))
-            .groupBy(course.id, course.name, course.level, course.distance,
-                course.estimatedTime, course.sigun, course.image,
-                challenge.status, challenge.endedAt,
-                rewardEvent.id, rewardEvent.status, rewardEvent.type, rewardEvent.value)
-            .orderBy(sortOption.toOrderSpecifier())
-            .offset(pageable.getOffset())
-            .limit(pageable.getPageSize())
-            .fetch();
-
-        JPAQuery<Long> total = queryFactory
-            .select(course.countDistinct())
-            .from(course)
-            .innerJoin(challenge.course, course)
-            .where(course.isDeleted.eq(false)
-                .and(challenge.status.eq(ChallengeStatus.ONGOING)));
 
         return PageableExecutionUtils.getPage(content, pageable, total::fetchOne);
     }
